Priority Queue/Linked-List structure like hugging face for tokenizer

+++ src/generator.rs
@@ -1,4 +1,3 @@
-// src/generator.rs
 use burn::tensor::backend::Backend;
 use burn::tensor::{Tensor, Int, ElementConversion};
 use crate::model::QuantCB;
@@ -16,15 +15,14 @@ impl TextGenerator {
         temperature: f32,
         rep_penalty: f32, 
     ) -> String {
-        let mut tokens = tokenizer.encode(prompt);
+        let mut tokens = tokenizer.encode(prompt); // This is Vec<u32>
         
         for _ in 0..max_tokens {
             let context_len = tokens.len();
             
-            // --- FIX 1: Convert usize tokens to Backend IntElem ---
             let data_vec: Vec<B::IntElem> = tokens
                 .iter()
-                .map(|&t| (t as i64).elem()) // Convert usize -> i64 -> Backend Element
+                .map(|&t| (t as i64).elem())
                 .collect();
 
             let input_tensor = Tensor::<B, 1, Int>::from_data(
@@ -32,22 +30,18 @@ impl TextGenerator {
                 device,
             ).unsqueeze::<2>(); // Shape: [1, seq_len]
 
-            // --- FIX 2: Dynamically size dummy targets to match input sequence length ---
-            // Using zeros() ensures valid vocab indices in case MTP calculates loss internally
             let dummy: Tensor<B, 2, Int> = Tensor::zeros([1, context_len], device);
             
-            // Forward pass using MTP logic
             let (logits, _, _, _, _, _, _) = model.forward_mtp(input_tensor, dummy, None);
             let [_, seq_len, vocab_size] = logits.dims();
             
-            // Extract logits for the last token in the sequence
             let last_logits_tensor = logits
                 .slice([0..1, (seq_len - 1)..seq_len, 0..vocab_size])
                 .reshape([vocab_size]);
                 
             let mut last_logits = last_logits_tensor.into_data().value;
 
-            // --- REPETITION PENALTY ---
+            // Repetition Penalty
             let window_size = usize::min(context_len, 64);
             for i in (context_len - window_size)..context_len {
                 let tok_idx = tokens[i] as usize;
@@ -61,7 +55,7 @@ impl TextGenerator {
                 }
             }
 
-            // --- TEMPERATURE & SAMPLING (Greedy w/ Temp) ---
+            // Greedy Sampling
             let mut max_val = f32::NEG_INFINITY;
             let mut next_token = 0;
             
@@ -70,14 +64,12 @@ impl TextGenerator {
                 
                 if adjusted_logit > max_val {
                     max_val = adjusted_logit;
-                    next_token = i;
+                    next_token = i; // i is usize
                 }
             }
             
-            tokens.push(next_token);
-
-            // Optional: Stop early if an end-of-text token is generated
-            // if next_token == EOT_TOKEN_ID { break; }
+            // FIX: Cast usize next_token to u32
+            tokens.push(next_token as u32);
         }
         
         tokenizer.decode(&tokens)
